Module with scanner auction working
