+++ a-core/src/test/java/core/service/BidServiceTests.java
@@ -4,10 +4,6 @@ import core.model.Bid;
 import core.model.Car;
 import core.model.Role;
 import core.model.User;
-import core.service.BidService;
-import core.service.CarService;
-import core.service.RoleService;
-import core.service.UserService;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
