--- car-scraper/src/main/java/core/scraper/manheim/ManheimMainPage.java
@@ -1,34 +1,0 @@
-package core.scraper.manheim;
-
-import core.scraper.BasePage;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-
-public class ManheimMainPage extends BasePage {
-    public ManheimMainPage(WebDriver webDriver) {
-        super(webDriver);
-    }
-
-    @FindBy(xpath = "//input[@placeholder='Username']")
-    private WebElement loginInput;
-
-    @FindBy(xpath = "//input[@placeholder='Password']")
-    private WebElement passwordInput;
-
-    //    @FindBy(xpath = "//*[@id='submit']")
-    @FindBy(id = "submit")
-    private WebElement signInButton;
-
-    public void enterLogin(String login) {
-        loginInput.sendKeys(login);
-    }
-
-    public void enterPassword(String password) {
-        passwordInput.sendKeys(password);
-    }
-
-    public void clickSignInButton() {
-        signInButton.click();
-    }
-}
