+++ car-scraper/src/main/resources/application.properties
@@ -1,0 +1,24 @@
+#hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+#hibernate.connection.driver_class=org.postgresql.Driver
+#hibernate.connection.url=jdbc:postgresql://localhost:5432/postgres
+#hibernate.connection.username=postgres
+#hibernate.connection.password=root
+#hibernate.show_sql=true
+#hibernate.hbm2ddl.auto=updat
+#
+datasource.driver-class-name = org.postgresql.Driver
+#hibernate.connection.driver_class=org.postgresql.Driver
+spring.sql.init.platform=postgres
+spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+spring.datasource.username=postgres
+spring.datasource.password=root
+
+spring.jpa.database=POSTGRESQL
+spring.jpa.show-sql=true
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
+spring.sql.init.mode=always
+spring.jpa.open-in-view=false
+spring.jpa.defer-datasource-initialization=true
+
