+++ car-scraper/src/main/resources/domain.xml
@@ -1,0 +1,8 @@
+<datasource jndi-name="java:jboss/datasources/MyDS" pool-name="MyPool" enabled="true" use-java-context="true">
+    <connection-url>jdbc:postgresql://localhost:5432/postgres</connection-url>
+    <driver>postgresql</driver>
+    <security>
+        <user-name>postgres</user-name>
+        <password>root</password>
+    </security>
+</datasource>
